[SONARPLUGINS-791] fix NPE when cleaning ASP projects before even building once

+++ maven/maven-dotnet-plugin/src/main/java/org/apache/maven/dotnet/CleanMojo.java
@@ -100,7 +100,9 @@ public class CleanMojo extends AbstractDotNetBuildMojo {
           = webProject.getWebPrecompilationDirectory();
         
         try {
-          FileUtils.cleanDirectory(precompilationDirectory);
+          if (precompilationDirectory!=null && precompilationDirectory.exists()) {
+            FileUtils.cleanDirectory(precompilationDirectory);
+          }
         } catch (IOException e) {
           throw new MojoExecutionException("error while cleaning web project "
               + visualStudioProject, e);
